Move reorder-related stuffs in visitor mod to reorder mod
